+++ packages/fe-fpm-writer/src/page/common.ts
@@ -15,37 +15,72 @@ import type { Manifest } from '../common/types';
 import { FCL_ROUTER } from '../common/defaults';
 
 /**
- * Add a new route to the provided route array, and update existing routes if necessary (e.g. if targets are defined as arrays for FCL).
- *
- * @param routes existing application routes (from the manifest)
- * @param config configuration object
+ * Suffix for patterns to support arbitrary paramters
  */
-export function updateRoutes(
+export const PATTERN_SUFFIX = ':?query:';
+
+/**
+ * Generates the pattern for a new route based on the input.
+ *
+ * @param routes existing routes
+ * @param targetEntity entity of the target
+ * @param nav navigation object
+ * @returns the generated pattern as string
+ */
+export function generateRoutePattern(
     routes: ManifestNamespace.Route[],
-    config: InternalCustomPage | InternalObjectPage | InternalListReport
-) {
-    const newRoute: ManifestNamespace.Route = {
-        name: `${config.entity}${config.name}`
-    };
-    if (config.navigation) {
-        const sourceRoute = routes.find((route) => route.name === config.navigation?.sourcePage);
-        const pattern = {
-            base: sourceRoute?.pattern?.replace(':?query:', ''),
-            navEntity: config.navigation.navEntity,
-            navKey: config.navigation.navKey ? `({${config.navigation.navEntity}Key})` : ''
-        };
-        newRoute.pattern = `${pattern.base}/${pattern.navEntity}${pattern.navKey}:?query:`;
-        // FCL only supports 3 columns, therefore, show the page in fullscreen if it is the 4th level of navigation
-        if (sourceRoute?.target?.constructor === Array && sourceRoute.target.length < 3) {
-            newRoute.target = [...sourceRoute.target, newRoute.name] as [string | ManifestNamespace.RouteTargetObject];
+    targetEntity: string,
+    nav?: Navigation
+): string {
+    const parts: string[] = [];
+    if (nav?.sourcePage && nav.navEntity) {
+        const sourceRoute = routes.find((route) => route.name === nav.sourcePage);
+        if (sourceRoute?.pattern) {
+            const basePattern = sourceRoute.pattern.replace(PATTERN_SUFFIX, '');
+            if (basePattern) {
+                parts.push(basePattern);
+                parts.push('/');
+                parts.push(nav.navEntity);
+            } else {
+                parts.push(targetEntity);
+            }
         } else {
-            newRoute.target = config.fcl ? [newRoute.name] : newRoute.name;
+            throw new Error('Navigation source invalid');
         }
-    } else {
-        newRoute.pattern = routes.length > 0 ? `${config.entity}:?query:` : ':?query:';
-        newRoute.target = config.fcl ? [newRoute.name] : newRoute.name;
+    } else if (routes.length > 0) {
+        parts.push(targetEntity);
     }
-    routes.push(newRoute);
+
+    if (nav?.navKey) {
+        parts.push(`({${nav?.navEntity ?? targetEntity}Key})`);
+    }
+    parts.push(PATTERN_SUFFIX);
+    return parts.join('');
+}
+
+/**
+ * Generates the target property for a route based on exiting routes, configurations and the target's name.
+ *
+ * @param routes existing routes
+ * @param name name of the target page
+ * @param fcl optional flag if FCL is enabled
+ * @param nav navigation object
+ * @returns the target property of a route
+ */
+export function generateRouteTarget(
+    routes: ManifestNamespace.Route[],
+    name: string,
+    fcl?: boolean,
+    nav?: Navigation
+): string | [string] | [ManifestNamespace.RouteTargetObject] {
+    if (nav?.sourcePage) {
+        const sourceRoute = routes.find((route) => route.name === nav.sourcePage);
+        // FCL only supports 3 columns, therefore, show the page in fullscreen if it is the 4th level of navigation
+        if (fcl && sourceRoute?.target?.constructor === Array && sourceRoute.target.length < 3) {
+            return [...sourceRoute.target, name] as [string];
+        }
+    }
+    return fcl ? [name] : name;
 }
 
 /**
@@ -63,7 +98,12 @@ export function getManifestJsonExtensionHelper(
                 value.routes = value.routes ?? [];
                 break;
             case 'routes':
-                updateRoutes(value as ManifestNamespace.Route[], config);
+                const routes = value as ManifestNamespace.Route[];
+                routes.push({
+                    name: `${config.entity}${config.name}`,
+                    pattern: generateRoutePattern(routes, config.entity, config.navigation),
+                    target: generateRouteTarget(routes, `${config.entity}${config.name}`, config.fcl, config.navigation)
+                });
                 break;
             default:
                 break;
@@ -110,8 +150,8 @@ export function validatePageConfig(basePath: string, config: CustomPage | Object
     validateBasePath(basePath, fs);
 
     // validate config against the manifest
-    const manifest = fs.readJSON(join(basePath, 'webapp/manifest.json')) as Manifest;
-    if (config.navigation) {
+    if (config.navigation?.sourcePage) {
+        const manifest = fs.readJSON(join(basePath, 'webapp/manifest.json')) as Manifest;
         if (!manifest['sap.ui5']?.routing?.targets?.[config.navigation.sourcePage]) {
             throw new Error(`Could not find navigation source ${config.navigation.sourcePage}!`);
         }
@@ -124,11 +164,9 @@ export function validatePageConfig(basePath: string, config: CustomPage | Object
             Object.assign(routes, manifest['sap.ui5']?.routing?.routes ?? {});
         }
 
-        const route = routes[config.navigation?.sourcePage];
+        const route = routes[config.navigation.sourcePage];
         if (!route || !route.pattern || !route.target) {
-            throw new Error(
-                `Missing or invalid routing configuration for navigation source ${config.navigation.sourcePage}!`
-            );
+            throw new Error(`Invalid routing configuration for navigation source ${config.navigation.sourcePage}!`);
         }
     }